Transport-project/requirements.txt
Django==2.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
